+++ src/lib.rs
@@ -68,8 +68,8 @@ mod analysis;
 mod message;
 
 /// Number of samples processed per second (Hertz) for a single channel of audio
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub struct SampleRate(pub u32);
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub struct SampleRate(pub f32);
 
 /// Input/output with this index does not exist
 #[derive(Debug, Clone, Copy)]
